Attempt resolution of SIGSEGV on RPi
Found source of weird bug generating the remote/undefined file. It is due
to reading in the configuration parameters for display in
UI_Remote_Config. Initially, there is no selected configuration and
apparently my implementation of RsyncUtils::readConfig touches the file
anyway.

As for the SIGSEGV, I'm trying changing some pointer/address/reference
whatever garbage in the .h/.cpp sources.