Add camera utility class
+ CameraUtils
+ ::start_preview, ::stop_preview
+ ::start_recording, ::stop_recording
+ ::capture
+ ::set_projectDir
+ ::get_preview_arg

Created a class to handle camera events for the RasPi. Was having lots
of trouble using the HQ camera... which is the entire point of this
project. The application works more or less as expected with USB camera,
but the HQ just causes tons of weird behavior when trying to use
QCamera. Instead, I am moving to raspistill/raspivid, and just using
tons of wrapped system calls in C++ which are then registered through to
QML. Feels a little dirty, but y'know.