Implement results of RPi deployment test
When testing deployment to the RPi, a few bugs of varying severity
were found. This commit contains those "hotfix" type changes which
were necessary to restore functionality. They're minor enough that
looking at the diffs for this commit is sufficient. Honestly, no
one will ever read this but me anyway. If you happen to be reading
this, please send me an email and tell me why. Because that's sort
of weird, honestly. :-)