Finalize RasPi camera implementation
Raspivid and raspistill should now be "fully" implemented, insofar as
being able to use them to take single images, videos of arbitrary
length, and frames for timelapses which are then automagically stitched
together. In the future of course I will expand the options available to
the user in terms of commands sent to raspistill/vid.