Test and fix camera behaviors on Pi
Fixed issues with timelapses not correctly taking frames, and not
correctly stitching them afterwards. Both fixed. Fixed issue with
videos not correctly starting and stopping. I think that is good
now -- but my check was mostly superficial. All in all, I think
the state of things now is fairly good as far as actual media
CAPTURE goes. Next up will be working on the external triggering,
improving the documentation, working on the remote sync more,
making the display of system info less horrible, and then we shall
go from there.