Modify camera_utils to fix recording bug
Recording was screwed up. Was recording when it claimed to be idle
and was idle when it claimed to be recording. Weird phasing thing
just caused by my uncertainty in whether raspivid starts recording
or paused. I found the -i flag in its documentation which allows
you to specify whether you want it to start paused or recording,
so I specified paused and that fixed the issue.