Resolve GPIO<->QML interface bugs
A few minor bugs with interfacing. Nothing that is really worth
my explaining here, the diffs should make it fairly obvious.
Other than one point, which is fairly interesting to the
extent that I might even save a note about it elsewhere:
when registering C++ classes as QML objects, and wrapping it in
a QObject, you cannot contain any other #include directive ASIDE
from the #include <QObject>. If you need others, you need to do
so in the .cpp source. I'm not sure why this is, or necessarily
if it is even real, but it seemed to completely resolve the
bizarre behavior.