Add C++ bones for additional cam params
C++ skeleton for adding things like annotations to images and frames in
timelapses, and setting user-defined encoding rates.

Also added an additional timer to update and display to the user the
remaining time until the next timed trigger, and reduced image
resolution in timelapse frames for the sake of reducing the absolutely-
ridiculous encoding times on a Pi.