Add remote storage support
Works... sort of. The program itself works fine, mostly. There is that
annoying behavior where sshfs seems to sometimes just hang forever. I
have no idea what causes it, and it is driving me seriously insane --
especially because I can find only one reference to similar behavior
online.

When it does work, though, there does seem to be one bug which occurs
truly in-program: the subfolders don't seem to be detected by
UI_Project. That is to say, if we create a remote storage location, and
then go to the projects context, nothing is visible. If we create a new
project, everything works mostly normally (if I check the location in
the filesystem manually, the folder is created correctly and where
expected), but for some reason nothing is visible in the folder list
model. This remote nonsense is super frustrating.