Implement remaining camera abilities
Added timed triggering for videos, as well as external triggering for
both videos and images. Timed triggering for videos allows for control
of both the length of the videos, and the empty time between them.
External video triggering will toggle the video recording state on any
edge, either rising or falling.

I also disabled the remote sync context for the time being. SSHFS may
simply be a no-go, it's too frustrating to deal with its bugs. Will
eventually migrate to NFS.

? Virtual keyboard is STILL not functional on Pi, although works fine on
desktop environment. May be forced to write my own.