Add noise resistance for edges
Modified the GPIOUtils::await_edge function in an attempt to reduce
spurious triggers.

Before, the loop would run in its thread as quickly as possible, and it
took only a single flip to trigger. Now, the loop samples the pin at 10
kHz, pushes the new value to the top of a stack, and pops a value off
the bottom. If all readings are the same, then we check if there was a
flip from the previous time when all values in the stack were equal. If
so, we update the new prev value, and check if the flip was in the proper
direction to constitute the user's selected edge condition.